useFolderLinkCount
